+++ code/cell.ui.sys/package.json
@@ -7,15 +7,22 @@
   "scripts": {
     "start": "parcel src/html/entry.html",
     "watch": "yarn start",
+    "test": "ts test",
+    "lint": "ts lint",
+    "build": "ts build",
+    "prepare": "ts prepare",
     "bundle": "sh sh/bundle.sh",
     "clean": "sh sh/clean.sh",
     "serve": "cd dist && serve"
   },
   "dependencies": {
     "@platform/cell.client": "0.2.70",
+    "@platform/cell.types": "^0.4.84",
     "@platform/polyfill": "0.0.1"
   },
   "devDependencies": {
+    "@platform/test": "0.1.7",
+    "@platform/ts": "3.7.20",
     "parcel-bundler": "1.12.4"
   },
   "files": [
