+++ code/cell.ui.sys/src/html/entry.html
@@ -11,6 +11,6 @@
     <div id="root">
       <div class="loading">Loading...</div>
     </div>
-    <script src="./entry.html.tsx"></script>
+    <script src="./entry.tsx"></script>
   </body>
 </html>
