--- code/cell.ui.sys/src/loader/index.ts
@@ -1,1 +1,0 @@
-export * from './loader';
