--- code/cell.ui.sys/src/loader/loader.ts
@@ -1,6 +1,0 @@
-import { init } from './loader.env';
-
-export const loader = {
-  init,
-  // root: () => import('../components/Root'),
-};
