+++ code/cell.ui.sys/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@platform/ts/tsconfig.json",
+  "include": ["src"],
+  "compilerOptions": {
+    "outDir": "lib",
+
+    "noUnusedLocals": false
+  }
+}
