+++ code/cell.ui.sys/tslint.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@platform/ts/tslint.json",
+  "rules": {
+    "no-console": false
+  }
+}
